Conflicts:
	app/views/admin/users/_table.html.slim
	app/views/admin/users/edit.html.slim

+++ app/controllers/admin/users_controller.rb
@@ -14,23 +14,23 @@ module Admin
 
     def create
       user.save
-      respond_with(:admin, user)
+      respond_with :admin, user, location: -> { edit_admin_user_path(user) }
     end
 
     def update
       user.save
-      respond_with(:admin, user)
+      respond_with :admin, user, location: -> { edit_admin_user_path(user) }
     end
 
     def destroy
-      user.destroy
-      respond_with(:admin, user)
+      user.destroy unless current_admin_user == user
+      respond_with :admin, user, location: -> { admin_users_path }
     end
 
     private
 
     def user_params
-      params.require(:user).permit(:full_name, :email, :password)
+      params.require(:user).permit(:full_name, :email, :password, :password_confirmation)
     end
   end
 end
